+++ Plugin~/WebRTCPlugin/Codec/CMakeLists.txt
@@ -1,8 +1,9 @@
 target_sources(WebRTCLib
   PRIVATE
+    CreateVideoCodecFactory.cpp
+    CreateVideoCodecFactory.h
     H264ProfileLevelId.cpp
     H264ProfileLevelId.h
-    VideoCodecImpl.h
 )
 
 if(Windows OR Linux)
