+++ Plugin~/WebRTCPlugin/UnityVideoDecoderFactory.cpp
@@ -3,22 +3,11 @@
 #include <media/engine/internal_decoder_factory.h>
 #include <modules/video_coding/include/video_error_codes.h>
 
+#include "Codec/CreateVideoCodecFactory.h"
 #include "GraphicsDevice/GraphicsUtility.h"
 #include "ProfilerMarkerFactory.h"
 #include "ScopedProfiler.h"
 #include "UnityVideoDecoderFactory.h"
-
-#if CUDA_PLATFORM
-#include "Codec/NvCodec/NvCodec.h"
-#endif
-
-#if UNITY_OSX || UNITY_IOS
-#import <sdk/objc/components/video_codec/RTCDefaultVideoDecoderFactory.h>
-#import <sdk/objc/native/api/video_decoder_factory.h>
-#elif UNITY_ANDROID
-#include "Android/AndroidCodecFactoryHelper.h"
-#include "Android/Jni.h"
-#endif
 
 namespace unity
 {
@@ -81,55 +70,34 @@ namespace webrtc
         std::unique_ptr<const ScopedProfilerThread> profilerThread_;
     };
 
-    static webrtc::VideoDecoderFactory* CreateNativeDecoderFactory(IGraphicsDevice* gfxDevice, ProfilerMarkerFactory* profiler)
-    {
-#if UNITY_OSX || UNITY_IOS
-        return webrtc::ObjCToNativeVideoDecoderFactory([[RTCDefaultVideoDecoderFactory alloc] init]).release();
-#elif UNITY_ANDROID
-        if (IsVMInitialized())
-            return CreateAndroidDecoderFactory().release();
-#elif CUDA_PLATFORM
-        if (gfxDevice->IsCudaSupport())
-        {
-            CUcontext context = gfxDevice->GetCUcontext();
-            return new NvDecoderFactory(context, profiler);
-        }
-#endif
-        return nullptr;
-    }
-
     UnityVideoDecoderFactory::UnityVideoDecoderFactory(IGraphicsDevice* gfxDevice, ProfilerMarkerFactory* profiler)
         : profiler_(profiler)
-        , internal_decoder_factory_(new webrtc::InternalDecoderFactory())
-        , native_decoder_factory_(CreateNativeDecoderFactory(gfxDevice, profiler))
+        , factories_()
     {
+        const std::vector<std::string> arrayImpl = {
+            kInternalImpl, kNvCodecImpl, kAndroidMediaCodecImpl, kVideoToolboxImpl
+        };
+
+        for (auto impl : arrayImpl)
+        {
+            auto factory = CreateVideoDecoderFactory(impl, gfxDevice, profiler);
+            if (factory)
+                factories_.emplace(impl, factory);
+        }
     }
 
     UnityVideoDecoderFactory::~UnityVideoDecoderFactory() = default;
 
     std::vector<webrtc::SdpVideoFormat> UnityVideoDecoderFactory::GetSupportedFormats() const
     {
-        std::vector<SdpVideoFormat> supported_codecs;
-        if (native_decoder_factory_)
-            for (const webrtc::SdpVideoFormat& format : native_decoder_factory_->GetSupportedFormats())
-                supported_codecs.push_back(format);
-        for (const webrtc::SdpVideoFormat& format : internal_decoder_factory_->GetSupportedFormats())
-            supported_codecs.push_back(format);
-        return supported_codecs;
+        return GetSupportedFormatsInFactories(factories_);
     }
 
     std::unique_ptr<webrtc::VideoDecoder>
     UnityVideoDecoderFactory::CreateVideoDecoder(const webrtc::SdpVideoFormat& format)
     {
-        std::unique_ptr<webrtc::VideoDecoder> decoder;
-        if (native_decoder_factory_ && format.IsCodecInList(native_decoder_factory_->GetSupportedFormats()))
-        {
-            decoder = native_decoder_factory_->CreateVideoDecoder(format);
-        }
-        else
-        {
-            decoder = internal_decoder_factory_->CreateVideoDecoder(format);
-        }
+        VideoDecoderFactory* factory = FindCodecFactory(factories_, format);
+        auto decoder = factory->CreateVideoDecoder(format);
         if (!profiler_)
             return decoder;
